\name{plot_moo}
\alias{plot_objectives}
\alias{plot_parameters}
\title{plot_moo}
\description{plot parameters and objectives for nsga2NREL}
\usage{
plot_objectives(objectives_save, objDim)
compplot_objectives(obj1, obj2, objDim)
plot_parameters(parameters_save, noVars)
}
\arguments{
  \item{objectives_save}{Number of objective functions}
  \item{parameters_save}{Number of objective functions}
  \item{objDim}{Number of objective functions}
  \item{noVars}{Number of objective functions}
  \item{obj1}{obj2}{objDim}
}

\author{
Brian Ball \email{brian.ball@nrel.gov}
}
\references{ http://www.r-bloggers.com/generate-uuids-in-r/
}

