\name{spea2NREL-package}
\alias{spea2NREL-package}
\docType{package}
\title{
A Parallel Strength Pareto Evolutionary Algorithm II based on R
}
\description{
A fast and elitist multiobjective genetic algorithm based on R. Modified spea2R by NREL for parallel F evals and mixed continuous and discrete variables.
}
\details{
\tabular{ll}{
Package: \tab spea2NREL\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2013-11-21\cr
License: \tab LGPL-3\cr
}
This package provide functions for box-constrained multiobjective optimization using a parallel Strength Pareto Evolutionary Algorithm II. Fast and Elitist sorting, crowding distance, tournament selection, simulated binary crossover, and polynomial mutation are called in the main program, spea2NREL, to complete the search.
}
\author{
Ching-Shih (Vince) Tsou <cstsou@mail.ntcb.edu.tw>

Maintainer: Brian Ball <Brian.Ball@nlr.gov>
}
\references{
Zitzler, E., Laumanns, M., Thiele, L., SPEA2: Improving the strength pareto 
  evolutionary algorithm. Technical report 103, Computer Engineering and Networks
  Laboratory (TIK), Swiss Federal Institute of Technology (ETH) Zurich, Gloriastrasse
  35, CH-8092 Zurich, Switzerland, 2001.
}

