\name{strengthRawFitness}
\alias{strengthRawFitness}
\title{
Strength Value and Raw Fitness
}
\description{
This function compute the strength value and raw fitness.
}
\usage{
strengthRawFitness(inputData)
}
\arguments{
  \item{inputData}{matrix of solutions with objective function values.}
}
\value{
return a matrix of strength value and raw fitness of all solutions.
}
\references{
Zitzler, E., Laumanns, M., Thiele, L., SPEA2: Improving the strength pareto evolutionary algorithm. Technical report 103, Computer Engineering and Networks Laboratory (TIK), Swiss Federal Institute of Technology (ETH) Zurich, Gloriastrasse 35, CH-8092 Zurich, Switzerland, 2001.
}
\author{
Ching-Shih (Vince) Tsou \email{cstsou@mail.ntcb.edu.tw}
}
