\name{ZTDfunctions}
\alias{ZDT1}
\alias{ZDT2}
\alias{ZDT3}
\alias{ZDT4}
\title{Test Problems for Multi-objective Optimization}
\description{Collection of functions implementing various spea2 test problems.}
\usage{
ZDT1(x)
ZDT2(x)
ZDT3(x)
ZDT4(x)
}
\arguments{
  \item{x}{Input vector.}
}
\value{
  Function value.
}
\author{
Ching-Shih (Vince) Tsou \email{cstsou@mail.ntcb.edu.tw}
}
