\name{nsga2NREL-package}
\alias{nsga2NREL-package}
\docType{package}
\title{
A Parallel Elitist Non-dominated Sorting Genetic Algorithm based on R
}
\description{
A fast and elitist multiobjective genetic algorithm based on R. Modified nsga2R by NREL for parallel F evals and mixed continuous and discrete variables.}
\details{
\tabular{ll}{
Package: \tab nsga2NREL\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2013-11-21\cr
License: \tab LGPL-3\cr
}
This package provide functions for box-constrained multiobjective optimization using a parallel elitist non-dominated sorting genetic algorithm - NSGA-II. Fast non-dominated sorting, crowding distance, tournament selection, simulated binary crossover, and polynomial mutation are called in the main program, nsga2NREL, to complete the search.
}
\author{
Ching-Shih (Vince) Tsou <cstsou@mail.ntcb.edu.tw>

Maintainer: Brian Ball <Brian.Ball@nrel.gov>
}
\references{
Deb, K., Pratap, A., Agarwal, S., and Meyarivan, T. (2002), " A fast and elitist multiobjective genetic algorithm: NSGA-II", IEEE Transactions on Evolutionary Computation, 6(2), 182-197.
}

