\name{plot_moo}
\alias{plot_objectives}
\alias{plot_obj}
\alias{plot_parameters}
\title{plot_moo}
\description{plot parameters and objectives for nsga2NREL}
\usage{
plot_objectives(objectives_save, objDim, gen)
plot_obj(objectives_save, objDim, gen)
plot_parameters(parameters_save, noVars, gen)
}
\arguments{
  \item{objectives_save}{Number of objective functions}
  \item{parameters_save}{Number of objective functions}
  \item{objDim}{Number of objective functions}
  \item{noVars}{Number of objective functions}
  \item{gen}{Number of objective functions}
}

\author{
Brian Ball \email{brian.ball@nrel.gov}
}
\references{ http://www.r-bloggers.com/generate-uuids-in-r/
}

