\name{plotMOO}
\alias{plotObjectives}
\alias{plotObjectivesGIF}
\alias{plotParameters}
\alias{compPlotObjectives}
\alias{compPlotObjectivesGIF}
\title{plotMOO}
\description{plot parameters and objectives for nsga2NREL}
\usage{
plotObjectives(objectivesSave, objDim)
plotObjectivesGIF(objectivesSave, objDim)
compPlotObjectives(obj1, obj2, objDim)
plotParameters(parametersSave, noVars)
compPlotObjectivesGIF(obj1, obj2, objDim)
}
\arguments{
  \item{objectivesSave}{Matrix of objective functions}
  \item{parametersSave}{Matrix of objective functions}
  \item{objDim}{Number of objective functions}
  \item{noVars}{Number of objective functions}
  \item{obj1}{Matrix of objective functions}
  \item{obj2}{Matrix of objective functions}
}

\author{
Brian Ball \email{brian.ball@nrel.gov}
}
\references{ http://www.r-bloggers.com/generate-uuids-in-r/
}

